+++ src/desktop/conversation_sidebar/conversation_sidebar.py
@@ -15,6 +15,7 @@ from desktop.conversation_sidebar.conversation_sidebar_tree_delegate import Conv
 from desktop.conversation_sidebar.conversation_sidebar_tree_view import ConversationSidebarTreeView
 from desktop.conversation_sidebar.conversation_sidebar_dag_model import ConversationSidebarDAGModel
 from desktop.conversation_sidebar.conversation_sidebar_index import ConversationSidebarIndex
+from desktop.file_utils import is_binary_image_file
 from desktop.language.language_manager import LanguageManager
 from desktop.message_box import MessageBox, MessageBoxButton, MessageBoxType
 from desktop.mindspace.mindspace_manager import MindspaceManager
@@ -914,6 +915,7 @@ class ConversationSidebar(SidebarBase):
             else:
                 # File context menu
                 edit_action = menu.addAction(strings.open_in_editor)
+                edit_action.setEnabled(not is_binary_image_file(path))
                 edit_action.triggered.connect(lambda: self._handle_edit_file(path))
                 preview_view_action = menu.addAction(strings.open_in_preview)
                 preview_view_action.triggered.connect(lambda: self._handle_preview_view_file(path))